refactor: moving constants around
