+++ crates/ironboy_core/Cargo.toml
@@ -7,8 +7,6 @@ edition = "2024"
 [dependencies]
 ironboy_cpu = {path = "../ironboy_cpu"}
 ironboy_system_bus = {path= "../ironboy_system_bus"}
-ironboy_apu = {path = "../ironboy_apu"}
-ironboy_ppu = {path = "../ironboy_ppu"}
 ironboy_cartridge = {path = "../ironboy_cartridge"}
 ironboy_joypad = {path = "../ironboy_joypad"}
 ironboy_common = {path = "../ironboy_common"}
