+++ crates/ironboy_core/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod gb;
 
-pub use ironboy_apu::{AUDIO_BUFFER_THRESHOLD, SAMPLING_FREQUENCY, SAMPLING_RATE};
+pub use ironboy_common::constants::{AUDIO_BUFFER_THRESHOLD, FPS, SAMPLING_FREQUENCY, SAMPLING_RATE, VIEWPORT_HEIGHT, VIEWPORT_WIDTH};
 pub use ironboy_joypad::JoypadButton;
-pub use ironboy_ppu::{FPS, VIEWPORT_HEIGHT, VIEWPORT_WIDTH};
