+++ crates/ironboy_cpu/src/instructions/branch.rs
@@ -1,6 +1,6 @@
 use crate::Cpu;
 
-use ironboy_common::MemoryInterface;
+use ironboy_common::memory::MemoryInterface;
 
 use super::Condition;
