+++ crates/ironboy_cpu/src/operands.rs
@@ -1,6 +1,6 @@
 use std::fmt;
 
-use ironboy_common::MemoryInterface;
+use ironboy_common::memory::MemoryInterface;
 
 use super::Cpu;
